fix: critical issues in adapter layer and test coverage
Blockers fixed:
- STM32 adapter: Fixed int-to-pointer cast (uintptr_t)
- UART validation: Added flush callback NULL check + test
- Session API: Removed dead code (session.transport field)
- Documentation: Fixed false claims in PROTOCOL_UART and PLATFORM_STATUS

Improvements:
- loxboot_uart_run_session now validates ctx->transport.flush != NULL
- Added test_uart_null_flush_rejected to prevent NULL pointer crashes
- Removed redundant session.transport from loxboot_uart_session_t struct
- Updated full UART update flow test with actual assertions
- Fixed CRC16 final XOR documentation (0x0000 not 0xFFFF)
- Corrected slot erase timing docs (first WRITE not session init)

Test results:
- 366 tests passing (362 core + 4 new validation tests)
- STM32 stub build now passes with cast fix
- GCC core-only verified
- All UART session validation tests passing

Known limitations documented:
- Erase granularity mismatch (core 52 bytes vs adapter sector alignment)
- Hardware validation still required
- Full power-loss testing not yet implemented

+++ adapters/stm32/loxboot_flash_stm32.c
@@ -16,7 +16,7 @@ static loxboot_err_t stm32_flash_read(void *ctx, uint32_t addr, uint8_t *buf, si
     }
 
     /* Internal flash is memory-mapped — direct pointer read */
-    const uint8_t *src = (const uint8_t *)addr;
+    const uint8_t *src = (const uint8_t *)(uintptr_t)addr;
     memcpy(buf, src, len);
     return LOXBOOT_OK;
 }
